refactor: share config assembly between factory and typegen
Extract the config list construction from rsEslint into an
assembleConfigs helper so scripts/typegen.ts reuses the same
code instead of hand-writing a parallel set of config calls.

The generated rule types are unchanged apart from the format
plugin's package-json/sort rule now being included.
